+++ Application/Feature/Folder/Commands/CreateFolderCommand.cs
@@ -19,7 +19,7 @@ namespace Application.Feature.Project.Folder.Commands
         public Task<bool> Handle(UpdateFolderCommand request, CancellationToken ct)
             => service.UpdateAsync(request.Id, request.Dto, request.UserId, request.DepartmentId, ct);
     }
-    public sealed record NewOrderFolderCommand(Guid Id, double NewOrder) : IRequest<bool>;
+    public sealed record NewOrderFolderCommand(Guid Id, int NewOrder) : IRequest<bool>;
 
     public sealed class NewOrderFolderCommandHandler(IFolderService service)
         : IRequestHandler<NewOrderFolderCommand, bool>
