+++ ProjectManagement.Client.Shared/Repositories/Calculation/ITaskRepository.cs
@@ -9,7 +9,7 @@ namespace ProjectManagement.Client.Shared.Repositories.Calculation
     public interface ITaskRepository
     {
         Task<List<ListOrderDTO>> GetAsync(int? id = null);
-        Task<bool> ReOrderAsync(int Id, double newOrder);
+        Task<bool> ReOrderAsync(int Id, int newOrder);
         Task<bool> CreateAsync(List<TaskPostDTO> model, int NetCalc);
         Task<List<TaskListMVVM>> GetByFilterAsync(FilterCalculationItemsDto offer);
         Task<bool> UpdateAsync(TaskPostDTO model, int id);
