+++ ProjectManagement.Client.Shared/Repositories/Calculation/Implement/CalculationRepository.cs
@@ -25,7 +25,7 @@ namespace ProjectManagement.Client.Shared.Repositories.Calculation.Implement
 
             return await _httpRepository.GetAsync<GetProjectCalcConfigDTO>(CalcURLBase + $"config/{m}/{con}/{com}/{t}/{st}");
         }
-        public async Task<bool> ReOrderAsync(int Id, double newOrder)
+        public async Task<bool> ReOrderAsync(int Id, int newOrder)
         {
             return await _httpRepository.GetAsync<bool>(CalcURLBase + URLConst.Calculation.ReOrder + $"/{Id}/{newOrder}");
         }
