+++ ProjectManagement.Client.Shared/Repositories/Project/IProjectRepository.cs
@@ -15,7 +15,7 @@ namespace ProjectManagement.Client.Shared.Repositories.Project
         Task<GetProjectCalcConfigDTO> GetConfig(int? m, int? con, int? com, int? t);
         Task<ProjectDetailsDTO> DetailsAsync(Guid projectId);
         Task<PostProjectDTO> GetToPostAsync(Guid projectId);
-        Task<bool> ReOrderAsync(Guid Id, double newOrder);
+        Task<bool> ReOrderAsync(Guid Id, int newOrder);
         Task<Guid> CreateAsync(PostProjectDTO create);
         Task<bool> UpdateAsync(Guid id, PostProjectDTO project);
         Task<bool> DeleteAsync(Guid id);
