+++ ProjectManagement.Shared/DTO/Calculation/CalculationDTO.cs
@@ -106,7 +106,7 @@ namespace ProjectManagement.Shared.DTO.Calculation
         public DateTime TenderQA { get; set; } = DateTime.Now;
         public DateTime StartDate { get; set; } = DateTime.Now;
         public DateTime EndDate { get; set; } = DateTime.Now.AddMonths(2);
-        public double Order { get; set; }
+        public int Order { get; set; }
         public DateTime? PublicationDate { get; set; } = DateTime.Now;
         public DateTime? DecisionDate { get; set; } = DateTime.Now;
         public List<HourlyPriceListGroupDTO> HourlyPrice { get; set; } = [];
@@ -180,7 +180,7 @@ namespace ProjectManagement.Shared.DTO.Calculation
     }
     public class ListCalculationDTO
     {
-        public double Order { get; set; }
+        public int Order { get; set; }
         public int Id { get; set; }
         public bool IsPrivate { get; set; }
         public string Name { get; set; } = string.Empty;
