+++ ProjectManagement/ProjectManagement/Controllers/v1/Project/CalculationsController.cs
@@ -21,7 +21,7 @@ namespace ProjectManagement.Server.Controllers.v1.Project
         }
         [Authorize(Roles = Tenant.AdminManger)]
         [HttpGet(URLConst.ReOrder + "/{Id}/{NewOrder}")]
-        public async Task<IActionResult> ReOrder(int Id, double NewOrder)
+        public async Task<IActionResult> ReOrder(int Id, int NewOrder)
         {
             return Ok(await MicroBus.Send(new NewOrderCalculationCommand(Id, NewOrder)));
         }
