+++ ProjectManagement/ProjectManagement/Controllers/v1/Project/ProjectsController.cs
@@ -55,7 +55,7 @@ namespace ProjectManagement.Server.Controllers.v1
         }
         [Authorize(Roles = Tenant.AdminManger)]
         [HttpGet(URLConst.ReOrder + "/{Id}/{NewOrder}")]
-        public async Task<IActionResult> ReOrder(Guid Id, double NewOrder)
+        public async Task<IActionResult> ReOrder(Guid Id, int NewOrder)
         {
             return Ok(await MicroBus.Send(new NewOrderProjectCommand(Id, NewOrder )));
         }
