+++ TaskResourceBlueprints/Entities/Resources/ResourceDefinition.cs
@@ -22,7 +22,7 @@ namespace TaskResourceBlueprints.Entities
 
         public ResourceTypesEnum ResType { get; set; }
         public string Name { get; set; } = null!;
-        [Range(0, double.MaxValue)] public double SortOrder { get; set; }
+        [Range(0, double.MaxValue)] public int SortOrder { get; set; }
         private ResourceMetadata _data = new();
         public ResourceMetadata Data
         {
